Implementadas funciones controlador

+++ src/Domain/Interfaces/INube.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.concurrent.Future;
 
 public interface INube extends Runnable{
-    Future<Nube> crearNubePuntos(int cantidad, int media, int desviacion);
+    Future<Nube> crearNubePuntos(int cantidad, double media, double desviacion);
     Future<Nube> crearNubePuntosRandom (int lower, int upper);
     ArrayList<Punto[]> splitX(int cantidad, int boundary);
 }
